Move path-exists.ts to a more generic non-ui path
