+++ app/src/lib/editors/launch.ts
@@ -1,5 +1,5 @@
 import { spawn, SpawnOptions } from 'child_process'
-import { pathExists } from '../../ui/lib/path-exists'
+import { pathExists } from '../path-exists'
 import { ExternalEditorError, FoundEditor } from './shared'
 import {
   expandTargetPathArgument,
