+++ app/src/lib/git/cherry-pick.ts
@@ -25,7 +25,7 @@ import { getCommit } from '.'
 import { IMultiCommitOperationProgress } from '../../models/progress'
 import { join } from 'path'
 import { readFile } from 'fs/promises'
-import { pathExists } from '../../ui/lib/path-exists'
+import { pathExists } from '../path-exists'
 
 /** The app-specific results from attempting to cherry pick commits*/
 export enum CherryPickResult {
