+++ app/src/lib/path-exists.ts
