+++ app/src/lib/shells/linux.ts
@@ -1,7 +1,7 @@
 import { spawn, ChildProcess } from 'child_process'
 import { assertNever } from '../fatal-error'
 import { parseEnumValue } from '../enum'
-import { pathExists } from '../../ui/lib/path-exists'
+import { pathExists } from '../path-exists'
 import { FoundShell } from './shared'
 import {
   expandTargetPathArgument,
